refactor: Move TestsBatchCompatibilityTagField to pants.core.goals.test
